+++ packages/money-account-utils/src/index.ts
@@ -23,8 +23,10 @@ export {
   getSharesForWithdrawal,
 } from './transactions.js';
 export {
+  MONEY_ACCOUNT_PREMIUM_VAULT_CONFIG_FLAG_NAME,
   MONEY_ACCOUNT_VAULT_CONFIG_FLAG_NAME,
   areMoneyAccountVaultConfigsEqual,
+  getMoneyAccountPremiumVaultConfig,
   getMoneyAccountVaultConfig,
   parseMoneyAccountVaultConfig,
 } from './vault-config.js';